# contracts/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "OpenDispatch",
    "description": "The REST half of the OpenDispatch API contract. SignalR board events and the offline-sync payloads are not describable here; they live in @opendispatch/contracts.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/customers": {
      "post": {
        "tags": [
          "Customers"
        ],
        "operationId": "CreateCustomer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerSummaryResponse"
                }
              }
            }
          },
          "default": {
            "description": "An error. Every failure this API returns — validation, not found, conflict, unauthorized, or unhandled — is a ProblemDetails body.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "get": {
        "tags": [
          "Customers"
        ],
        "operationId": "ListCustomers",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerPageResponse"
                }
              }
            }
          },
          "default": {
            "description": "An error. Every failure this API returns — validation, not found, conflict, unauthorized, or unhandled — is a ProblemDetails body.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/customers/{id}": {
      "get": {
        "tags": [
          "Customers"
        ],
        "operationId": "GetCustomer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerResponse"
                }
              }
            }
          },
          "default": {
            "description": "An error. Every failure this API returns — validation, not found, conflict, unauthorized, or unhandled — is a ProblemDetails body.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "put": {
        "tags": [
          "Customers"
        ],
        "operationId": "UpdateCustomer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "description": "An error. Every failure this API returns — validation, not found, conflict, unauthorized, or unhandled — is a ProblemDetails body.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/customers/{id}/locations": {
      "post": {
        "tags": [
          "Customers"
        ],
        "operationId": "AddServiceLocation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceLocationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceLocationResponse"
                }
              }
            }
          },
          "default": {
            "description": "An error. Every failure this API returns — validation, not found, conflict, unauthorized, or unhandled — is a ProblemDetails body.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/customers/{id}/locations/{locationId}": {
      "put": {
        "tags": [
          "Customers"
        ],
        "operationId": "UpdateServiceLocation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "locationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceLocationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "description": "An error. Every failure this API returns — validation, not found, conflict, unauthorized, or unhandled — is a ProblemDetails body.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Customers"
        ],
        "operationId": "RemoveServiceLocation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "locationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "description": "An error. Every failure this API returns — validation, not found, conflict, unauthorized, or unhandled — is a ProblemDetails body.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateCustomerRequest": {
        "required": [
          "name",
          "email",
          "phone"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Their name, personal or trading."
          },
          "email": {
            "type": [
              "null",
              "string"
            ],
            "description": "An email address, or `null` if there isn't one."
          },
          "phone": {
            "type": [
              "null",
              "string"
            ],
            "description": "A phone number, or `null` if there isn't one."
          }
        },
        "description": "The body of `POST /customers`."
      },
...
      "CustomerPageResponse": {
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomerSummaryResponse"
            },
            "description": "The customers on this page, by name."
          },
          "total": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "How many the tenant has altogether.",
            "format": "int32"
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Which page this is, counting from one.",
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "How many this page was allowed to hold.",
            "format": "int32"
          }
        },
        "description": "One page of customers. The body of `GET /customers?page={n}&amp;pageSize={n}`."
      },
      "CustomerResponse": {
        "required": [
          "id",
          "name",
          "email",
          "phone",
          "locations"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Their identity.",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Their name, personal or trading."
          },
          "email": {
            "type": [
              "null",
              "string"
            ],
            "description": "Their email address, or `null` if there isn't one."
          },
          "phone": {
            "type": [
              "null",
              "string"
            ],
            "description": "Their phone number, or `null` if there isn't one."
          },
          "locations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServiceLocationResponse"
            },
            "description": "The places they want work done, in the order they were added."
          }
        },
        "description": "One customer, as the clients see them: who they are, how to reach them, and where they want work done."
      },
      "CustomerSummaryResponse": {
        "required": [
          "id",
          "name",
          "email",
          "phone"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Their identity, which is what a caller picks them by.",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Their name, personal or trading."
          },
          "email": {
            "type": [
              "null",
              "string"
            ],
            "description": "Their email address, or `null` if there isn't one."
          },
          "phone": {
            "type": [
              "null",
              "string"
            ],
            "description": "Their phone number, or `null` if there isn't one."
          }
        },
        "description": "A customer as `GET /customers` lists them: enough to recognise and to reach, and nothing else."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ServiceLocationRequest": {
        "required": [
          "label",
          "address",
          "latitude",
          "longitude"
        ],
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "description": "What the customer calls it — \"Home\", \"Unit 4\", \"the Croydon branch\"."
          },
          "address": {
            "type": "string",
            "description": "The postal address a technician would be given."
          },
          "latitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Where it is, in decimal degrees between -90 and 90.",
            "format": "double"
          },
          "longitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Where it is, in decimal degrees between -180 and 180.",
            "format": "double"
          }
        },
        "description": "The body of `POST /customers/{id}/locations` and `PUT .../locations/{locationId}`."
      },
      "ServiceLocationResponse": {
        "required": [
          "id",
          "label",
          "address",
          "latitude",
          "longitude"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The identity a job points at.",
            "format": "uuid"
          },
          "label": {
            "type": "string",
            "description": "What the customer calls it."
          },
          "address": {
            "type": "string",
            "description": "The postal address a technician would be given."
          },
          "latitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Where it is, in decimal degrees.",
            "format": "double"
          },
          "longitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Where it is, in decimal degrees.",
            "format": "double"
          }
        },
        "description": "One of a customer's service locations, as the clients see it."
      },
...
      "UpdateCustomerRequest": {
        "required": [
          "name",
          "email",
          "phone"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Their name, personal or trading."
          },
          "email": {
            "type": [
              "null",
              "string"
            ],
            "description": "An email address, or `null` if there isn't one."
          },
          "phone": {
            "type": [
              "null",
              "string"
            ],
            "description": "A phone number, or `null` if there isn't one."
          }
        },
        "description": "The body of `PUT /customers/{id}`."
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "The token POST /auth/login returns, sent as \"Authorization: Bearer <token>\".",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}