# src/Api/Customers/CustomerEndpoints.cs
using MediatR;
using OpenDispatch.Api.Auth;
using OpenDispatch.Api.ErrorHandling;
using OpenDispatch.Application.Customers.AddServiceLocation;
using OpenDispatch.Application.Customers.CreateCustomer;
using OpenDispatch.Application.Customers.GetCustomer;
using OpenDispatch.Application.Customers.ListCustomers;
using OpenDispatch.Application.Customers.RemoveServiceLocation;
using OpenDispatch.Application.Customers.UpdateCustomer;
using OpenDispatch.Application.Customers.UpdateServiceLocation;
using OpenDispatch.Contracts.Customers;
using OpenDispatch.Domain.Identifiers;

namespace OpenDispatch.Api.Customers;

/// <summary>
/// Customers and the service locations they own — CRUD over the slices steps 32 and 47 built
/// (Document 3, step 47).
/// </summary>
/// <remarks>
/// Every route here is <see cref="AuthPolicies.AdminOrDispatcher"/>: this is the office-side
/// surface a dispatcher uses to take on a customer or look one up before booking a job, and an
/// admin uses for the same reasons. A technician reaches customer data, if at all, through the
/// sync endpoints (step 50) — never through here.
/// </remarks>
public static class CustomerEndpoints
{
    public static IEndpointRouteBuilder MapCustomerEndpoints(this IEndpointRouteBuilder endpoints)
    {
        var customers = endpoints.MapGroup("/customers")
            .RequireAuthorization(AuthPolicies.AdminOrDispatcher)
            .WithTags("Customers");

        customers.MapPost("/", CreateAsync).WithName("CreateCustomer")
            .Produces<CustomerSummaryResponse>(StatusCodes.Status201Created);
        customers.MapGet("/", ListAsync).WithName("ListCustomers")
            .Produces<CustomerPageResponse>();
        customers.MapGet("/{id:guid}", GetAsync).WithName("GetCustomer")
            .Produces<CustomerResponse>();
        customers.MapPut("/{id:guid}", UpdateAsync).WithName("UpdateCustomer")
            .Produces(StatusCodes.Status204NoContent);

        customers.MapPost("/{id:guid}/locations", AddLocationAsync).WithName("AddServiceLocation")
            .Produces<ServiceLocationResponse>(StatusCodes.Status201Created);
        customers.MapPut("/{id:guid}/locations/{locationId:guid}", UpdateLocationAsync)
            .WithName("UpdateServiceLocation")
            .Produces(StatusCodes.Status204NoContent);
        customers.MapDelete("/{id:guid}/locations/{locationId:guid}", RemoveLocationAsync)
            .WithName("RemoveServiceLocation")
            .Produces(StatusCodes.Status204NoContent);

        return endpoints;
    }

    private static async Task<IResult> CreateAsync(
        CreateCustomerRequest request,
        ISender sender,
        CancellationToken cancellationToken)
    {
        var result = await sender
            .Send(new CreateCustomerCommand(request.Name, request.Email, request.Phone), cancellationToken)
            .ConfigureAwait(false);

        return result.ToHttpResult(id => Results.Created(
            $"/customers/{id.Value}",
            new CustomerSummaryResponse(id.Value, request.Name, request.Email, request.Phone)));
    }

    /// <remarks>
    /// Both parameters are optional and both have defaults, so a caller that has never heard of
    /// paging still gets an answer — a first page, rather than the whole customer book this used to
    /// hand over.
    /// </remarks>
    private static async Task<IResult> ListAsync(
        ISender sender,
        CancellationToken cancellationToken,
        int? page = null,
        int? pageSize = null)
    {
        var query = new ListCustomersQuery(
            page ?? 1,
            pageSize ?? ListCustomersQuery.DefaultPageSize);

        var result = await sender.Send(query, cancellationToken).ConfigureAwait(false);

        return result.ToHttpResult(customers => Results.Ok(new CustomerPageResponse(
            [.. customers.Items.Select(ToSummary)],
            customers.Total,
            customers.Page,
            customers.PageSize)));
    }

    private static async Task<IResult> GetAsync(Guid id, ISender sender, CancellationToken cancellationToken)
    {
        var result = await sender
            .Send(new GetCustomerQuery(CustomerId.From(id)), cancellationToken)
            .ConfigureAwait(false);

        return result.ToHttpResult(customer => Results.Ok(ToResponse(customer)));
    }

    private static async Task<IResult> UpdateAsync(
        Guid id,
        UpdateCustomerRequest request,
        ISender sender,
        CancellationToken cancellationToken)
    {
        var command = new UpdateCustomerCommand(CustomerId.From(id), request.Name, request.Email, request.Phone);
        var result = await sender.Send(command, cancellationToken).ConfigureAwait(false);

        return result.ToHttpResult();
    }

    private static async Task<IResult> AddLocationAsync(
        Guid id,
        ServiceLocationRequest request,
        ISender sender,
        CancellationToken cancellationToken)
    {
        var command = new AddServiceLocationCommand(
            CustomerId.From(id), request.Label, request.Address, request.Latitude, request.Longitude);
        var result = await sender.Send(command, cancellationToken).ConfigureAwait(false);

        return result.ToHttpResult(locationId => Results.Created(
            $"/customers/{id}/locations/{locationId.Value}",
            new ServiceLocationResponse(locationId.Value, request.Label, request.Address, request.Latitude, request.Longitude)));
    }

    private static async Task<IResult> UpdateLocationAsync(
        Guid id,
        Guid locationId,
        ServiceLocationRequest request,
        ISender sender,
        CancellationToken cancellationToken)
    {
        var command = new UpdateServiceLocationCommand(
            CustomerId.From(id),
            ServiceLocationId.From(locationId),
            request.Label,
            request.Address,
            request.Latitude,
            request.Longitude);
        var result = await sender.Send(command, cancellationToken).ConfigureAwait(false);

        return result.ToHttpResult();
    }

    private static async Task<IResult> RemoveLocationAsync(
        Guid id,
        Guid locationId,
        ISender sender,
        CancellationToken cancellationToken)
    {
        var command = new RemoveServiceLocationCommand(CustomerId.From(id), ServiceLocationId.From(locationId));
        var result = await sender.Send(command, cancellationToken).ConfigureAwait(false);

        return result.ToHttpResult();
    }

    private static CustomerSummaryResponse ToSummary(CustomerSummary customer) =>
        new(customer.Id.Value, customer.Name, customer.Email, customer.Phone);

    private static CustomerResponse ToResponse(CustomerDetail customer) => new(
        customer.Id.Value,
        customer.Name,
        customer.Email,
        customer.Phone,
        [.. customer.Locations.Select(location =>
            new ServiceLocationResponse(location.Id.Value, location.Label, location.Address, location.Latitude, location.Longitude))]);
}